// src/tarefas/enums/task.enum.ts
export enum TaskEnums{
    TASK_NOT_CREATED = "Não foi possível criar a tarefa. Verifique os parâmetros e tente novamente.",
    TASK_NOT_FOUND = "Não foi possível encontrar tarefas por meio dos parâmetros informados.",
    TASK_NOT_UPDATED = "Não foi possível atualizar a tarefa. Verifique os campos informados e tente novamente.",
    TASK_NOT_BETWEEN_DATES = "Não foram encontradas tarefas dentro do período informado.",
    ASSOCIATED_USER_NOT_FOUND = "Não foram encontradas tarefas para esse usuário."
}